+++ test/unit/parse.js
@@ -24,6 +24,7 @@ describe("parse - get mode", () => {
       manifestUrl: "https://nwjs.io/versions",
       cache: true,
       ffmpeg: false,
+      logLevel: "info"
     };
 
     deepStrictEqual(actualOptions, expectedOptions);
@@ -51,6 +52,7 @@ describe("parse - get mode", () => {
         "https://raw.githubusercontent.com/nwutils/nw-builder/main/src/util/osx.arm.versions.json",
       cache: true,
       ffmpeg: false,
+      logLevel: "info"
     };
 
     deepStrictEqual(actualOptions, expectedOptions);
@@ -73,6 +75,7 @@ describe("parse - get mode", () => {
       manifestUrl: "https://nwjs.io/versions",
       cache: true,
       ffmpeg: false,
+      logLevel: "info"
     };
 
     deepStrictEqual(actualOptions, expectedOptions);
@@ -95,6 +98,7 @@ describe("parse - get mode", () => {
       manifestUrl: "https://nwjs.io/versions",
       cache: true,
       ffmpeg: false,
+      logLevel: "info"
     };
 
     deepStrictEqual(actualOptions, expectedOptions);
@@ -117,6 +121,7 @@ describe("parse - get mode", () => {
       manifestUrl: "https://nwjs.io/versions",
       cache: true,
       ffmpeg: true,
+      logLevel: "info"
     };
 
     deepStrictEqual(actualOptions, expectedOptions);